separates mysqrt()
