+++ codes/mysqrt/mysqrt/mysqrt.vcxproj.filters
@@ -18,5 +18,13 @@
     <ClCompile Include="mysqrt.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="mymath.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="mymath.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
